requirements.txt
bleach==3.2.1
Flask==1.1.2
flask-login==0.5.0
GitPython==3.1.11
Requests==2.25.0
simplejson==3.17.2
pyYAML==5.3.1
google-auth==1.24.0
